MPS-25809 AWT events are not allowed inside write action: java.awt.event.InvocationEvent[INVOCATION_DEFAULT,runnable=Delay: -58ms; class com.intellij.util.concurrency.SchedulingWrapper$MyScheduledFutureTask (future done),notifier=null,catchExceptions=false,when=1492330841239] on sun.awt.X11.XToolkit@341d572c
Caused by my change from #invokeAndWait to #invokeLater. Supposedly the
change was a mistake. I assume that the declared model conflict resolution
workflow is correct: when we save and reload
the conflicting model from the disk and only after that we ask user whether
he wants to replace it with his model (backed up) or leave as it was on
the disk.
Also removing a double check in the #checkConflictOnSave -- I cannot see
the way it can help to get rid of the model inconsistency if it is
possible at all.