Introduce EditorMenuTrace and add it to the transformation and substitute menu context. EditorMenuTrace holds stack of EditorMenuTraceInfo. EditorMenuTraceInfo contains EditorMenuDescriptor and holds the references to the parent and children.
So while the process of creating the menu items list the tree of EditorMenuTraceInfo is build via context.getEditorMenuTrace().push/popTraceInfo().

Make all the parts of the process such as menu parts, implicit menus etc. add their nodes to the tree.

Add the method returning the trace info to the SubstituteAction, SubstituteMenuItem and TransformationMenuItem so we can show the trace