MPS-24997 Backspace should select relevant nodes for deletion
Introduce DeletionApprover interface and it's implementations. DeletionApproverImpl is used if the settings checkbox "two steps deletion" is checked, DummyDeletionApprover is used otherwise.
Before deleting the node client may now call deletionApprover.approveForDeletion(). DeletionApproverImpl will highlight the node set to be deleted.
DeletionApproverUtil is used for convenience: checking the node for null, whether is it already approved for deletion and whether the selection covers the whole node, so it doesn't need to be approved