MPS-24997 Backspace should select relevant nodes for deletion
Introduce ApproveDelete_Operation which client can use in custom DELETE and BACKSPACE actions.
It is quite similar to SelectInEditor_Operation, it should specify editor context and may specify cell id.

The result of this operation is true if the specified cell was (cell specified by id, or node's big cell) was succesfully approved for deletion (and highlighted). In this case custom DELETE and BACKSPACE action should stop.
Otherwise it should proceed with the deletion.

Update custom lang.editor DELETE and BACKSPACE actions using this operation