MPS-26231 All run configurations are broken after mps restart
quick unfix!
idea platform api gets changed again. The initialization of the run
configurations differs as well. Loosing any ability to read the before
run tasks (which persist correctly btw!) due to the lazy initialization of
the mapping beforeId -> beforeRunTaskProvider which early accesses the
extension point (which we use to register our custom before tasks) and
is not reloadable at all.
It is set at the first access of RunManager smashing all our hopes to
insert our custom plugin tasks there
since we are able to reload plugins only after the ProjectRunConfigurationsManager
is constructed (I chose #projectOpened, which also docs in the PRCM suggest, though you might try doing better)

In summary before tasks are set by default from the template
configurations for now.

PS
Also I am not sure that this early access in order to invoke #saveState
is absolutely necessary.

The righteous path is to ask IDEA to give a proper extension here.
Otherwise each platform change will be a huge pain for us (as it is now,
obviously).