Providing warning instead of error message during the generation of an EditorTestCase.
Currently I am fixing a particular generation problem in mbeddr, however
a generic scenario is as follows: EditorTestCase runtime requires the presence of
annotation in the original model while during the generation phase
(lang.test) it might be as well reduced. This is a consequence of interpretation
logic in the editor (typesystem as well) tests runtime.