MPS-24685 Behavior throws method not found instead of NPE
Now it just return the default value (see nullability test in the
BHTestRef root). Thing is the generated method call lacked the concrete
concept of the method declaration. I changed the generator of the
reflective invocation and now it always provides the concept. It lets us
to find the method with the method id and inside SMethodImpl it makes
all necessary calculations for returning the default values in the case
of null operand