Fix MPS-26951 Registered model factories can disappear from new model dialog
* 'factories' parameter is CopyOnWriteArrayList from ModelFactoryService
(ModelFactoryService.myCustomModelFactories)
* IterableUtil.asList() returns original list (not a copy) is the passed
iterable was indeed a list.
* Collections.reverse() works in-place

So, it happened the code in question was reversing in-place the collection
of factories passed to it.
When indexing happened (clearing caches is the way to reproduce) in
multiple threads simultaneously, several clients were reversing in-place
the same collection. Collections#reverse() works via List#set(), which
is thread-safe in CopyOnWriteArrayList(), but a sequence of set() calls is not.

So, it all ended with a collection of factories having the right number
of elements but duplicate entries, effectively loosing some of the factories.

With this change we copy the list of factories upon each query. Note,
in-place reversing of CopyOnWriteArrayList did even more copyings.