MPS-26755 Duplicate entries in the code completion menu MPS-26926 content assist menu contains double entries when using interface concept
Track menu lookups which are used within the same transformation/substitute context, so there will be no duplication of substitue/transformation items when the concept inherits from the interface in multiple ways
Add corresponding editor tests.
Such approach restricts user to include same menu twice to another menu: add tests for that too.
Also add tests that similar context is used to create items when the node and menu locations are not changed by the IncludeTransformationMenuTransformationMenuPartTest.
So we don't need to add all possible variants of changing the context in "include transformation menu part" in editor tests and just use "current node" and "current location"