lang.plugin: respect the case when MPS manages loading of app/project component parts itself.
When app/project components are loaded from MPS-controlled module (with the help of ModulePluginContributor), no need to expose their entries as extensions.
To tell whether it's MPS to control, I'm a bit harsh and check both compileInMPS and the fact we control loaded classes sensibly (not by naming convention, though for these there's no reason to have extension, too.).