MPS-27086 CustomActionsSchema saving erroneously hide some actions
In com.intellij.ide.ui.customization.CustomizableActionsPanel Idea uses groups names instead of Ids to show them in tree, and when it tries to optimise tree in method com.intellij.ide.ui.customization.CustomizationUtil#optimizeSchema, it fails on some MPS groups with similar titles.
Add postfix to groups as fix.

TypesystemActions_Plugin and TypesystemActions_Devkit also have to remove 'caption', which was hide (because of 'is popup' was false), but still used in generation.