MPS-23821: Handle submodules under module directory on rename
If module folder name matches module name, it will be renamed with module. In some rare cases module folder can contain other modules. For example it is default layout for language with sandbox or runtime solutions.

Updating module descriptor for submodules will ensure that all paths in it will be renamed and saved to module file.

Rename dialog will notify user about such submodules and mark ones, that will be renamed - if it contains main module name in it.