ModelIdentity: improve editing experience (completion).
This includes some hacks to deal with the fact editor could not override a menu for a grandchild node. Container for ModelIdentity knows what menu is expected, but ModelPointer's presentation is ModelName, and its substitute menu doesn't respect grandparent's setting.
To overcome, we delegate to parent's menu from ModelName, and provide contributions into ModelIdentity's default menu from another language (the one with the container concept). Unfortunatelly, this forces us to have 'extends' relation between the languages