MPS-26874 Addition of base placeholder is displayed badly in diff dialog
Fix the problems of interacting between vcs and child attributes.
Since child attributes depend on the nodes in the same role  (the order is important) we should treat them not as children in 'smodelAttribute" collection, but we should treat them together with the children in the same role.
So the idea is to change the semantic of the NodeGroupChange class which now represents the change of the group of the the children and child attributes in the same role.
So we chould change all the clients of this class accordingly.

Also add vcs tests for adding, removing child attributes, commenting and uncommenting