MPS-26114 Bad size of completion menu on first appearance
Setting the size of the popup based on the preferred size was leading to an error: when popup added the borders to the content, the size of the content became bigger than the popup and the scrollbar always showed up.
Change the logic: don't set the size directly, but provide the custom layout manager to the main panel.

Also do the following changes:
- init the ui only when the show() method is called
- init the list in NodeSubstituteChooser every time it's shown and set it to null when it's hidden. (that could cause a memory leak)
- set setSelectionIndex() and getSelectionIndex() to private: they were mistakenly public