DescriptorIO: facilitate use of custom macro helper in descriptor persistence.
Module descriptors now is the one responsible to deal with path macros. As long as there are few locations we need to read modules at (e.g. build language), we need a control over what macros are available at this time. Therefore, we provide an alternative way to supply MacroHelper in addition to a default one.