platform 181.3741: fix breaking of @NotNull in DefaultFileIndexFacade
In method DefaultFileIndexFacade#isInContent if it was initialized with default project call of method getBaseDir() returns null, what is violating contract of VfsUtilCore#isAncestor(@NotNull VirtualFile, @NotNull VirtualFile, boolean).
This happens in next test after vfs refresh, which send notification with MessageBus.

Initialization of default project happens in ProjectManagerImpl#newProject with call of checkProjectLeaksInTests() and in case of parameter useDefaultProjectSettings set to true.

Switch to MockFileIndexFacade for now to avoid problem and as better solution, than disable leak project check.