MPS-27101 Name of the model shall not end with '.'
It looks like the only way to get this and similar exception is a library without a name, so that created stub module has empty name, and its counterpart project module tries to get instantiated with 'module.@project_stub' name. Prevent creation of a library and notify user about the issue.