MPS-27643 IdeaFile doesnt work properly on Windows
So the problem appeared on windows after I fixated the immutability
contract on the IdeaFile. The equality relation had to rely on
the path strings given at the construction of IdeaFile. Thus some
IdeaFiles which point to the same location on disk still are not equal
to each other due to the minor difference in the path strings.
After a bit of thinking I decided to change the failing test. The reason
is simple -- the IdeaFile equality relation does not promise to be that
strong, i.e. to hold if and only if the target location of two IdeaFiles is the same.
If the client of IdeaFile wants to check such information he might address
the #getPath method instead.