Ant: Drop reflective cloneTo as it's of no use
There's a defect in cloneTo which assumed method return values are either java primitive types or come from bootstrap cl (relied on return type from source method).
Moreover, I can hardly imagine a scenario when Script class of the task and the one of the Worker are different; or there's a need to ensure Script class loaded by worker is identical to Script class supplied as its construction argument.