Expose classes of j.m.tool.common for MPS developers
It's part of com.intellij plugin and is contributed with mps-tool.jar in distribution.
The file is a mechanism to contribute classes of com.intellij plugin when running from sources.
As long as the module is not exposed as a stub, I've introduced a distinct 'unassigned' group, though an alternative, with all mps-tool stuff in a single group looks fruitful as well.