MPS-26523 replace dubious isUserModel() with !isStubModel() check.
isUserModel respects closed set of cases what to deem as 'user' model.
It's not apparent why a model with e.g. a custom stereotype is not a 'user' one, while our technical @descriptor is.
As for transient models, filtering thereof generally happens in the initial scope (e.g. project.getScope() shall not give any transient model).