test language: set focus correctly after executing "add test label annotation"
Previously the complete annotation was selecteded in the current editor
context this made it impossible to directly start typing after adding
the annotation. As the node annotation and the annontated node where
selected typing anything would cause MPS to override the node itself.

Added a attracts focus to the name cell of the TestNodeAnnotation and
changed the intention "Add Test Label Annotation" to select the first
editable cell in the current editor context.