MPS-27867 Problem with the contract of the EDTExecutorInternal.
It is impossible to stay concise with the IDEA write-safety model +
guarantee the FIFO principle of our queue + flushing of the queue regardless of
the current modality + allow to take the write lock within the queue
tasks.
It urges us to hack the IDEA write-safety model, namely the
ModelAccess#runWriteInEDT is vulnerable to the sudden model changes
(change of the file on disk).
Since we still do not have any protection for the MPS model, this
solution is tolerable for the minor release however the question remains to
be answered in an honest way.

There is an alternative solution advised by Alex Shatalin as follows.
We might have something like this in the #flushQueueLaterInEDT method
body:
<code>
  ApplicationManager.getApplication().invokeLater(() -> {
    ModalityState current = ModalityState.current(); // since we are in EDT we are able to find here the current modality
    ApplicationManager.getApplication().invokeLater(this::flushQueue, current);
  }, ModalityState.any())
</code>
However this double async invocation still does not guarantee us the
flushing of the queue regardless of the modal dialogs stack since a new
modal dialog might be opened before we get to our #flushQueue and the
calculated <code>current</code> modality would dominate the new
modality.
In order to minimize such a risk we might add a ModalityListener
which will check for such situations and reschedule the flush every time
a new modality appears.

This alternative solution was implemented and supposedly works (at least
in some common scenarios and of course in the case of the MPS-27867).

I am choosing the first solution since it requires less changes despite
the fact that in the former solution we use the knowledge of IDEA's
internals a bit, whilst in the latter we do not.