MPS-27918 can't place a cursor in the middle of the broken reference cell
Create two similar test: for the smart reference and for the regular reference.
Both tests replace reference target with the other node, go to the reference, put the caret to the center and replace the reference.

Also in order to get the correct reference while executing tests in the same process, restrict the scope of the reference.