MPS-27918 can't place a cursor in the middle of the broken reference cell
Correct the tests which remove the invalid references respecting that the error cells are now editable again.
The behaviour is not optimal since we need to press backspace/delete twice, but this is the behaviour a the long time ago.
This behaviour (necessity to press backspace/delete twice) should be changed.