DefaultModelRoot: replace direct instantiation with factory method and ModelRootDescriptor
ModelRoot is an object initialized from a descriptor, and the right way to construct a root is to prepare a descriptor, not vice versa.
OTOH, descriptor has no means to get configured without exposing implementation of DefaultModelRoot (i.e. memento keys). Therefore, use of a single method with access to DMR internals helps to keep this under control.